Control.Monad.Counter: Generalize IntCounter interface
MonadCounter is introduced which generalizes the interface of IntCounter so that
we can generate fresh variables in other monads without `lift`ing operators
repeatedly. The FreshVariable class generalizes the interface for generating
fresh variables so that each variable is not tied to a particular monad.

The fresh variable counter is changed from Int to Natural. The principled reason
to do this is that it is a counter, and Natural is the type of counting
numbers. The practical reasons are that Natural does not admit negative numbers
and will not overflow.