translatePredicate: Always attempt to translate builtin equality
`translatePredicate` attempts the following, in order, to translate the children
of an `\equals` predicate:

1. Translate children as builtin domain value patterns.
2. Translate children as predicates.
3. Translate the entire predicate uninterpreted.

We always attempt translation as builtin domain value patterns first because
that is most specific. We cannot use the sorts alone to choose between
translating as builtin domain value patterns or predicates, because predicates
can occur in any sort.