maybeEvaluatePattern: Do not add child conditions to side conditions
Adding child conditions to the side conditions without simplifying the result
leads to a denormalized side condition. It is not appropriate to call the
condition simplifier here, so we must not use the child conditions at all. The
child conditions will ultimately be added to the side conditions by the
pattern simplifier, leading to the re-evaluation of any equations which might
require the child conditions.