Add data type Variable1
The goal of refactoring is to replace type Variable and the wrappers around it
with type Variable1, and rename the latter. Note that this will make the
SortedVariable class obsolete because all variables will carry a sort field at
the top level; therefore, we will modify the SortedVariable class to carry the
intermediate definitions we need for refactoring, and remove the class in the
end.